+++ CHANGELOG.md
@@ -1,5 +1,18 @@
 # Change log
 
+## Version 0.20.0
+* Added connection pool class and made PocketWriter instances thread-safe (https://github.com/kaykay-dv/pocketsearch/issues/44)
+* Added possibility to use "legacy" (tables already defined in a sqlite3 database) tables (https://github.com/kaykay-dv/pocketsearch/issues/42)
+* Fixed a bug resulting in SQL errors when using the @ symbol at certain places (https://github.com/kaykay-dv/pocketsearch/issues/49)
+* Added proper handling for empty string queries (https://github.com/kaykay-dv/pocketsearch/issues/43)
+* Added possiblity to define own "id" fields (https://github.com/kaykay-dv/pocketsearch/issues/42)
+* Fixed a bug in SQL trigger definitions
+
+**Breaking changes**
+* PocketWriters are now using transactions. If an exception is raised the transaction is rolled back. (https://github.com/kaykay-dv/pocketsearch/issues/45)
+* Building the spell checker dictionary has now to be invoked explicitly (https://github.com/kaykay-dv/pocketsearch/issues/46)
+* Table creation (FTS5 tables) are now wrapped in an own transaction (https://github.com/kaykay-dv/pocketsearch/issues/45)
+
 ## Version 0.12.0
 * Added context managers PocketReader and PocketWriter
 * Added Spell Checker